Change value in MAF & Freq1 columns for 1 test SNP
In the test data set in the check/inputfiles directory all SNPs had
their Freq1 > MAF, so with the new 'flipmaf' code, all would be flipped.
In order to check if nothing changes for variants with Freq1 < MAF this
commit changes these values for one of the test SNPs.

This commit also updates the results in the checks/verified_results
directory accordingly so that the check_dose_input.sh and
check_probabel_chunk.sh scripts continue to function as before.