Add check scripts for the --flipmaf option
This implements the regular 'text genotype input' vs 'filevector
genotype input' checks.
NOTE: the 2df flipmaf vs flipmaf_fv check is currently disabled. The
value is off in the 6th significant digit, which would be ok in general,
but I'm using diff to check, so the check fails.

Added to these regular checks are checks to verify e.g. that beta ->
-beta for the additive model and beta_domin -> -beta_recess.

The Makefile.am file also includes dependencies for the flipmaf
checks (run the regular checks first so that those output files exist).