Add allelesFlipped column to output with --flipmaf
If the --flipmaf command line option is used, the output files will
contain an extra column called 'allelesFlipped' indicating whether the
reference and effect alleles have been flipped (1) or not (0).

To achieve this I added a std::vector<bool> to the mlinfo class. So this
class is no longer just a copy of the data from the .mlinfo file data.
If the --flipmaf option is set, the regdata::update_snp() and
coxph_data::update_snp() functions will set the appropriate entry in
this vector to true.